Move model structs and tests into models.rs
