+++ library/src/network.rs
@@ -1,7 +1,6 @@
 // This file's job is to deal with the update_server and network side
 // of the updater library.
 
-use serde::{Deserialize, Serialize};
 use std::fs::File;
 use std::io::Write;
 use std::path::Path;
@@ -9,6 +8,7 @@ use std::string::ToString;
 
 use crate::cache::UpdaterState;
 use crate::config::{current_arch, current_platform, UpdateConfig};
+use crate::models::{PatchCheckRequest, PatchCheckResponse};
 
 // https://stackoverflow.com/questions/67087597/is-it-possible-to-use-rusts-log-info-for-tests
 #[cfg(test)]
@@ -125,44 +125,6 @@ pub fn testing_set_network_hooks(
     });
 }
 
-#[derive(Debug, Deserialize)]
-pub struct Patch {
-    /// The patch number.  Starts at 1 for each new release and increases
-    /// monotonically.
-    pub number: usize,
-    /// The hex-encoded sha256 hash of the final uncompressed patch file.
-    /// Legacy: originally "#" before we implemented hash checks (remove).
-    pub hash: String,
-    /// The URL to download the patch file from.
-    pub download_url: String,
-}
-
-#[derive(Debug, Serialize)]
-pub struct PatchCheckRequest {
-    /// The Shorebird app_id built into the shorebird.yaml in the app.
-    pub app_id: String,
-    /// The Shorebird channel built into the shorebird.yaml in the app.
-    pub channel: String,
-    /// The release version from AndroidManifest.xml, Info.plist in the app.
-    pub release_version: String,
-    /// The latest patch number that the client has downloaded.
-    /// Not necessarily the one it's running (if some have been marked bad).
-    /// We could rename this to be more clear.    
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub patch_number: Option<usize>,
-    /// Platform (e.g. "android", "ios", "windows", "macos", "linux").
-    pub platform: String,
-    /// Architecture we're running (e.g. "aarch64", "x86", "x86_64").
-    pub arch: String,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct PatchCheckResponse {
-    pub patch_available: bool,
-    #[serde(default)]
-    pub patch: Option<Patch>,
-}
-
 pub fn send_patch_check_request(
     config: &UpdateConfig,
     state: &UpdaterState,
@@ -210,31 +172,6 @@ pub fn download_to_path(
 
 #[cfg(test)]
 mod tests {
-    use crate::network::PatchCheckResponse;
-
-    #[test]
-    fn check_patch_request_response_deserialization() {
-        let data = r###"
-    {
-        "patch_available": true,
-        "patch": {
-            "number": 1,
-            "download_url": "https://storage.googleapis.com/patch_artifacts/17a28ec1-00cf-452d-bdf9-dbb9acb78600/dlc.vmcode",
-            "hash": "#"
-        }
-    }"###;
-
-        let response: PatchCheckResponse = serde_json::from_str(data).unwrap();
-
-        assert!(response.patch_available == true);
-        assert!(response.patch.is_some());
-
-        let patch = response.patch.unwrap();
-        assert_eq!(patch.number, 1);
-        assert_eq!(patch.download_url, "https://storage.googleapis.com/patch_artifacts/17a28ec1-00cf-452d-bdf9-dbb9acb78600/dlc.vmcode");
-        assert_eq!(patch.hash, "#");
-    }
-
     // This confirms that the default network hooks throw an error in cfg(test).
     // In cfg(not(test)) they should be set to the default implementation
     // which makes real network calls.
